refactor(home): separate home data logic into useHomeBusinesses hook
- Introduce useHomeBusinesses to encapsulate React Query, Geoapify, connectivity, and filtering logic for the Home experience
- Update HomeScreen to focus on presentation and navigation, consuming home data exclusively via useHomeBusinesses
- Document the new hook and layering in README.md to clarify the architecture and separation of concerns
